+++ nextcloudappstore/core/scaffolding/archive.py
@@ -9,6 +9,7 @@ from django.template import Context
 from django.template import Template
 
 from nextcloudappstore.core.facades import resolve_file_relative_path
+from nextcloudappstore.settings.base import APP_SCAFFOLDING_PROFILES
 
 
 def build_files(args: Dict[str, str]) -> Dict[str, str]:
@@ -22,8 +23,11 @@ def build_files(args: Dict[str, str]) -> Dict[str, str]:
         'author_name': args['author_name'],
         'author_mail': args['author_email'],
         'author_homepage': args['author_homepage'],
+        'issue_tracker': args['issue_tracker'],
         'categories': args['categories'],
+        'nextcloud_version': platform
     }
+    vars.update(APP_SCAFFOLDING_PROFILES.get(platform, {}))
     relative_base = 'app-templates/%i/app/' % platform
     base = resolve_file_relative_path(__file__, relative_base)