+++ nextcloudappstore/core/scaffolding/forms.py
@@ -18,11 +18,11 @@ def get_categories():
 
 def get_versions():
     tpls = listdir(resolve_file_relative_path(__file__, 'app-templates'))
-    return [(v, v) for v in settings.PLATFORM_VERSIONS if v in tpls]
+    return [(v, v) for v in tpls]
 
 
 def validate_id(input: str) -> str:
-    regex = r'^([A-Z][a-z]+)+$'
+    regex = r'^([A-Z][a-z]*)+$'
     if not re.match(regex, input):
         raise ValidationError(_('The app name must be camel case e.g. MyApp'))
 
@@ -36,6 +36,8 @@ class AppScaffoldingForm(Form):
     author_name = CharField(max_length=80, label=_('Author\'s full name'))
     author_email = EmailField(label=_('Author\'s e-mail'))
     author_homepage = URLField(label=_('Author\'s homepage'), required=False)
+    issue_tracker = URLField(label=_('Issue tracker URL'), required=True,
+                             help_text=_('Bug reports and feature requests'))
     categories = MultipleChoiceField(required=True, label=_('Categories'),
                                      choices=lazy(get_categories, list),
                                      help_text=_('Hold down CTRL and click to '