+++ nextcloudappstore/urls.py
@@ -52,3 +52,11 @@ urlpatterns += i18n_patterns(
     url(r'feeds/releases.atom', AppReleaseAtomFeed(),
         name='feeds-releases-atom'),
 )
+
+if settings.DEBUG:
+    try:
+        import debug_toolbar
+
+        urlpatterns += [url(r'^__debug__/', include(debug_toolbar.urls)), ]
+    except ImportError:
+        pass