+++ scripts/maintenance/update.sh
@@ -34,10 +34,10 @@ pip install --upgrade wheel
 pip install --upgrade pip
 pip install -r requirements/base.txt
 pip install -r requirements/production.txt
-python manage.py compilemessages
-python manage.py collectstatic
 python manage.py migrate
 python manage.py loaddata nextcloudappstore/**/fixtures/*.json
+python manage.py collectstatic
+python manage.py compilemessages
 python manage.py importdbtranslations
 
 deactivate